Correctly describe possible sizes of packed GUIDs
The current docs refer to the size of the packedGUID being read as 64 bits. That is only true after unpacking it; the initial size can range from 2 to 9 bytes. This is of importance to ensure that users understand the size variation of packets that contain packed GUIDs and don't read past the byte buffer length.